geocrypt-backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Security & Authentication
pyotp==2.9.0
cryptography==41.0.7

# File Handling
Pillow==10.1.0

# CORS for React integration
django-cors-headers==4.2.0

# Environment variables
python-decouple==3.8

# Geo-location
geopy==2.4.0

# Date/Time
pytz==2023.3

# Development tools
django-extensions==3.2.3
django-filter==23.3

# AI/Monitoring
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
joblib==1.3.2

# HTTP requests
requests==2.31.0

# Async (optional for future)
channels==4.0.0
channels-redis==4.1.0
